Extract common collage utilities
