+++ src/components/collage/components/CanvasCollagePreview.js
@@ -3,398 +3,9 @@ import PropTypes from 'prop-types';
 import { Box, IconButton, Typography, Menu, MenuItem, ListItemIcon } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
 import { OpenWith, Check, Place, Crop, DragIndicator } from '@mui/icons-material';
-import { layoutDefinitions } from '../config/layouts';
+import { getBorderPixelSize, createLayoutConfig, detectBorderZones, parseGridToRects } from '../../../utils/collageLayout';
 import CaptionEditor from './CaptionEditor';
 
-
-
-/**
- * Helper function to get border pixel size based on percentage and component width
- */
-const getBorderPixelSize = (borderThickness, componentWidth = 400) => {
-  if (typeof borderThickness === 'number') {
-    return Math.round((borderThickness / 100) * componentWidth);
-  }
-  return 0;
-};
-
-/**
- * Helper function to create layout config from template
- */
-const createLayoutConfig = (template, panelCount) => {
-  if (!template) return null;
-  
-  try {
-    // Look up the original layout in the layout definitions
-    const panelCountKey = Math.max(2, Math.min(panelCount, 5));
-    const categories = layoutDefinitions[panelCountKey];
-    
-    // Search for the layout in all categories (wide, tall, square)
-    if (categories) {
-      const foundLayout = Object.keys(categories).reduce((result, category) => {
-        if (result) return result; // Already found a layout
-        
-        const layouts = categories[category];
-        const originalLayout = layouts.find(l => l.id === template.id);
-        
-        if (originalLayout && typeof originalLayout.getLayoutConfig === 'function') {
-          return originalLayout;
-        }
-        
-        return null;
-      }, null);
-      
-      if (foundLayout) {
-        return foundLayout.getLayoutConfig();
-      }
-    }
-    
-    // Fallback to a basic grid layout in case of error
-    return {
-      gridTemplateColumns: `repeat(${Math.ceil(Math.sqrt(panelCount))}, 1fr)`,
-      gridTemplateRows: `repeat(${Math.ceil(panelCount / Math.ceil(Math.sqrt(panelCount)))}, 1fr)`,
-      gridTemplateAreas: null,
-      items: Array(panelCount).fill({ gridArea: null })
-    };
-  } catch (error) {
-    console.error("Error creating layout config:", error, template);
-    // Fallback to a basic grid layout in case of error
-    return {
-      gridTemplateColumns: `repeat(${Math.ceil(Math.sqrt(panelCount))}, 1fr)`,
-      gridTemplateRows: `repeat(${Math.ceil(panelCount / Math.ceil(Math.sqrt(panelCount)))}, 1fr)`,
-      gridTemplateAreas: null,
-      items: Array(panelCount).fill({ gridArea: null })
-    };
-  }
-};
-
-/**
- * Helper function to parse CSS grid template areas string
- */
-const parseGridTemplateAreas = (gridTemplateAreas) => {
-  if (!gridTemplateAreas) return {};
-  
-  const areas = {};
-  
-  // Split by quotes to get individual rows
-  // Handle both 'row1' 'row2' and "row1" "row2" formats
-  const cleanString = gridTemplateAreas.trim();
-  let rows;
-  
-  if (cleanString.includes('" "')) {
-    // Format: "main main" "left right"
-    rows = cleanString.split('" "').map(row => 
-      row.replace(/"/g, '').trim().split(/\s+/)
-    );
-  } else if (cleanString.includes("' '")) {
-    // Format: 'main main' 'left right'
-    rows = cleanString.split("' '").map(row => 
-      row.replace(/'/g, '').trim().split(/\s+/)
-    );
-  } else {
-    // Single row or space-separated format
-    rows = [cleanString.replace(/['"]/g, '').trim().split(/\s+/)];
-  }
-  
-  // Find the bounds of each named area
-  rows.forEach((row, rowIndex) => {
-    row.forEach((areaName, colIndex) => {
-      if (areaName !== '.' && areaName !== '') {
-        if (!areas[areaName]) {
-          areas[areaName] = {
-            rowStart: rowIndex,
-            rowEnd: rowIndex,
-            colStart: colIndex,
-            colEnd: colIndex
-          };
-        } else {
-          // Extend the area bounds
-          areas[areaName].rowStart = Math.min(areas[areaName].rowStart, rowIndex);
-          areas[areaName].rowEnd = Math.max(areas[areaName].rowEnd, rowIndex);
-          areas[areaName].colStart = Math.min(areas[areaName].colStart, colIndex);
-          areas[areaName].colEnd = Math.max(areas[areaName].colEnd, colIndex);
-        }
-      }
-    });
-  });
-  
-  return areas;
-};
-
-/**
- * Helper function to detect draggable border zones
- */
-const detectBorderZones = (layoutConfig, containerWidth, containerHeight, borderPixels) => {
-  if (!layoutConfig) return [];
-  
-  const zones = [];
-  // Make threshold larger for mobile - minimum 16px hit area for better touch interaction
-  const threshold = Math.max(16, borderPixels * 2);
-  
-  // Parse grid columns and rows to get track sizes
-  let columnSizes = [1];
-  let rowSizes = [1];
-  
-  if (layoutConfig.gridTemplateColumns) {
-    if (layoutConfig.gridTemplateColumns.includes('repeat(')) {
-      const repeatMatch = layoutConfig.gridTemplateColumns.match(/repeat\((\d+),/);
-      if (repeatMatch) {
-        const count = parseInt(repeatMatch[1], 10);
-        columnSizes = Array(count).fill(1);
-      }
-    } else {
-      const frMatches = layoutConfig.gridTemplateColumns.match(/(\d*\.?\d*)fr/g);
-      if (frMatches) {
-        columnSizes = frMatches.map(match => {
-          const value = match.replace('fr', '');
-          return value === '' ? 1 : parseFloat(value);
-        });
-      }
-    }
-  }
-  
-  if (layoutConfig.gridTemplateRows) {
-    if (layoutConfig.gridTemplateRows.includes('repeat(')) {
-      const repeatMatch = layoutConfig.gridTemplateRows.match(/repeat\((\d+),/);
-      if (repeatMatch) {
-        const count = parseInt(repeatMatch[1], 10);
-        rowSizes = Array(count).fill(1);
-      }
-    } else {
-      const frMatches = layoutConfig.gridTemplateRows.match(/(\d*\.?\d*)fr/g);
-      if (frMatches) {
-        rowSizes = frMatches.map(match => {
-          const value = match.replace('fr', '');
-          return value === '' ? 1 : parseFloat(value);
-        });
-      }
-    }
-  }
-  
-  // Only create zones if we have multiple columns or rows
-  if (columnSizes.length <= 1 && rowSizes.length <= 1) {
-    return [];
-  }
-  
-  // Calculate positions of grid lines
-  const totalColumnFr = columnSizes.reduce((sum, size) => sum + size, 0);
-  const totalRowFr = rowSizes.reduce((sum, size) => sum + size, 0);
-  
-  const horizontalGaps = Math.max(0, columnSizes.length - 1) * borderPixels;
-  const verticalGaps = Math.max(0, rowSizes.length - 1) * borderPixels;
-  
-  const availableWidth = containerWidth - (borderPixels * 2) - horizontalGaps;
-  const availableHeight = containerHeight - (borderPixels * 2) - verticalGaps;
-  
-  const columnFrUnit = availableWidth / totalColumnFr;
-  const rowFrUnit = availableHeight / totalRowFr;
-  
-  // Create vertical border zones (between columns) - only if we have multiple columns
-  if (columnSizes.length > 1) {
-    let currentX = borderPixels;
-    for (let i = 0; i < columnSizes.length - 1; i += 1) {
-      currentX += columnSizes[i] * columnFrUnit;
-      
-      zones.push({
-        type: 'vertical',
-        index: i,
-        x: currentX - threshold / 2,
-        y: 0,
-        width: threshold,
-        height: containerHeight,
-        centerX: currentX,
-        cursor: 'col-resize',
-        id: `vertical-${i}` // Add ID for debugging
-      });
-      
-      currentX += borderPixels;
-    }
-  }
-  
-  // Create horizontal border zones (between rows) - only if we have multiple rows
-  if (rowSizes.length > 1) {
-    let currentY = borderPixels;
-    for (let i = 0; i < rowSizes.length - 1; i += 1) {
-      currentY += rowSizes[i] * rowFrUnit;
-      
-      zones.push({
-        type: 'horizontal',
-        index: i,
-        x: 0,
-        y: currentY - threshold / 2,
-        width: containerWidth,
-        height: threshold,
-        centerY: currentY,
-        cursor: 'row-resize',
-        id: `horizontal-${i}` // Add ID for debugging
-      });
-      
-      currentY += borderPixels;
-    }
-  }
-  
-  return zones;
-};
-
-/**
- * Helper function to parse CSS grid template and convert to panel rectangles
- */
-const parseGridToRects = (layoutConfig, containerWidth, containerHeight, panelCount, borderPixels) => {
-  const rects = [];
-  
-  // Calculate the available space (subtract borders)
-  const totalPadding = borderPixels * 2;
-  const availableWidth = containerWidth - totalPadding;
-  const availableHeight = containerHeight - totalPadding;
-  
-  // Parse grid template columns/rows to get exact dimensions and track sizes
-  let columns = 1;
-  let rows = 1;
-  let columnSizes = [1]; // Default: single column taking full width
-  let rowSizes = [1];    // Default: single row taking full height
-  
-  // Parse columns
-  if (layoutConfig.gridTemplateColumns) {
-    if (layoutConfig.gridTemplateColumns.includes('repeat(')) {
-      const repeatMatch = layoutConfig.gridTemplateColumns.match(/repeat\((\d+),/);
-      if (repeatMatch) {
-        columns = parseInt(repeatMatch[1], 10);
-        columnSizes = Array(columns).fill(1); // All equal size
-      }
-    } else {
-      // Parse individual fr units like "2fr 1fr" or "1fr 1fr 1fr"
-      const frMatches = layoutConfig.gridTemplateColumns.match(/(\d*\.?\d*)fr/g);
-      if (frMatches) {
-        columns = frMatches.length;
-        columnSizes = frMatches.map(match => {
-          const value = match.replace('fr', '');
-          return value === '' ? 1 : parseFloat(value);
-        });
-      }
-    }
-  }
-  
-  // Parse rows
-  if (layoutConfig.gridTemplateRows) {
-    if (layoutConfig.gridTemplateRows.includes('repeat(')) {
-      const repeatMatch = layoutConfig.gridTemplateRows.match(/repeat\((\d+),/);
-      if (repeatMatch) {
-        rows = parseInt(repeatMatch[1], 10);
-        rowSizes = Array(rows).fill(1); // All equal size
-      }
-    } else {
-      // Parse individual fr units like "2fr 1fr" or "1fr 1fr 1fr"
-      const frMatches = layoutConfig.gridTemplateRows.match(/(\d*\.?\d*)fr/g);
-      if (frMatches) {
-        rows = frMatches.length;
-        rowSizes = frMatches.map(match => {
-          const value = match.replace('fr', '');
-          return value === '' ? 1 : parseFloat(value);
-        });
-      }
-    }
-  }
-  
-  // Calculate total fractional units
-  const totalColumnFr = columnSizes.reduce((sum, size) => sum + size, 0);
-  const totalRowFr = rowSizes.reduce((sum, size) => sum + size, 0);
-  
-  // Calculate gaps - only between panels, not at edges
-  const horizontalGaps = Math.max(0, columns - 1) * borderPixels;
-  const verticalGaps = Math.max(0, rows - 1) * borderPixels;
-  
-  // Calculate base unit sizes
-  const columnFrUnit = (availableWidth - horizontalGaps) / totalColumnFr;
-  const rowFrUnit = (availableHeight - verticalGaps) / totalRowFr;
-  
-  // Helper function to calculate cumulative position and size for a cell
-  const getCellDimensions = (col, row) => {
-    // Calculate X position: sum of all previous column widths + gaps
-    let x = borderPixels;
-    for (let c = 0; c < col; c += 1) {
-      x += columnSizes[c] * columnFrUnit + borderPixels;
-    }
-    
-    // Calculate Y position: sum of all previous row heights + gaps  
-    let y = borderPixels;
-    for (let r = 0; r < row; r += 1) {
-      y += rowSizes[r] * rowFrUnit + borderPixels;
-    }
-    
-    // Calculate current cell dimensions
-    const width = columnSizes[col] * columnFrUnit;
-    const height = rowSizes[row] * rowFrUnit;
-    
-    return { x, y, width, height };
-  };
-
-  if (layoutConfig.areas && layoutConfig.areas.length > 0 && layoutConfig.gridTemplateAreas) {
-    // Use grid template areas - need to parse the actual grid areas
-    const gridAreas = parseGridTemplateAreas(layoutConfig.gridTemplateAreas);
-    
-    layoutConfig.areas.slice(0, panelCount).forEach((areaName, index) => {
-      const areaInfo = gridAreas[areaName];
-      if (areaInfo) {
-        // Calculate position and size based on grid area bounds
-        let x = borderPixels;
-        let y = borderPixels;
-        let width = 0;
-        let height = 0;
-        
-                 // Calculate X position and width
-         for (let c = 0; c < areaInfo.colStart; c += 1) {
-           x += columnSizes[c] * columnFrUnit + borderPixels;
-         }
-         for (let c = areaInfo.colStart; c <= areaInfo.colEnd; c += 1) {
-           width += columnSizes[c] * columnFrUnit;
-         }
-         // Add gaps between columns within the area
-         if (areaInfo.colEnd > areaInfo.colStart) {
-           width += (areaInfo.colEnd - areaInfo.colStart) * borderPixels;
-         }
-         
-         // Calculate Y position and height
-         for (let r = 0; r < areaInfo.rowStart; r += 1) {
-           y += rowSizes[r] * rowFrUnit + borderPixels;
-         }
-         for (let r = areaInfo.rowStart; r <= areaInfo.rowEnd; r += 1) {
-           height += rowSizes[r] * rowFrUnit;
-         }
-         // Add gaps between rows within the area
-         if (areaInfo.rowEnd > areaInfo.rowStart) {
-           height += (areaInfo.rowEnd - areaInfo.rowStart) * borderPixels;
-         }
-        
-        rects.push({
-          x,
-          y,
-          width,
-          height,
-          panelId: `panel-${index + 1}`,
-          index
-        });
-      }
-    });
-  } else {
-    // Use items array or simple grid
-    for (let i = 0; i < panelCount; i += 1) {
-      const col = i % columns;
-      const row = Math.floor(i / columns);
-      const { x, y, width, height } = getCellDimensions(col, row);
-      
-      rects.push({
-        x,
-        y,
-        width,
-        height,
-        panelId: `panel-${i + 1}`,
-        index: i
-      });
-    }
-  }
-  
-  return rects;
-};
 
 /**
  * Canvas-based Collage Preview Component
