+++ src/constants/collage.js
@@ -1,0 +1,19 @@
+export const BORDER_COLOR_PRESETS = [
+  { color: '#FFFFFF', name: 'White' },
+  { color: '#000000', name: 'Black' },
+  { color: '#FF0000', name: 'Red' },
+  { color: '#0000FF', name: 'Blue' },
+  { color: '#FFFF00', name: 'Yellow' },
+  { color: '#00FF00', name: 'Green' },
+];
+
+export const TEXT_COLOR_PRESETS = [
+  { color: '#ffffff', name: 'White' },
+  { color: '#000000', name: 'Black' },
+  { color: '#ff0000', name: 'Red' },
+  { color: '#00ff00', name: 'Green' },
+  { color: '#0000ff', name: 'Blue' },
+  { color: '#ffff00', name: 'Yellow' },
+  { color: '#ff00ff', name: 'Magenta' },
+  { color: '#00ffff', name: 'Cyan' },
+];
